Move 'apis/operator.marin3r' to 'apis/operator' otherwise the csv generation does not work correctly
