+++ apis/operator/v1alpha1/discoveryservice_types.go
