+++ config/manifests/bases/marin3r.clusterserviceversion.yaml
@@ -196,6 +196,99 @@ spec:
         displayName: Rtds Layer Resource Name
         path: envoyStaticConfig.rtdsLayerResourceName
       version: v1alpha1
+    - description: DiscoveryService represents an envoy discovery service server. Currently only one DiscoveryService per cluster is supported.
+      displayName: DiscoveryService
+      kind: DiscoveryService
+      name: discoveryservices.operator.marin3r.3scale.net
+      specDescriptors:
+      - description: ServiceConfig configures the way the DiscoveryService endpoints are exposed
+        displayName: Service Config
+        path: ServiceConfig
+      - displayName: Name
+        path: ServiceConfig.name
+      - displayName: Type
+        path: ServiceConfig.type
+      - description: Debug enables debugging log level for the discovery service controllers. It is safe to use since secret data is never shown in the logs.
+        displayName: Debug
+        path: debug
+      - description: Image holds the image to use for the discovery service Deployment
+        displayName: Image
+        path: image
+      - description: MetricsPort is the port where metrics are served. Defaults to 8383.
+        displayName: Metrics Port
+        path: metricsPort
+      - description: PKIConfig has configuration for the PKI that marin3r manages for the different certificates it requires
+        displayName: PKIConfig
+        path: pkiConfg
+      - displayName: Root Certificate Authority
+        path: pkiConfg.rootCertificateAuthority
+      - displayName: Duration
+        path: pkiConfg.rootCertificateAuthority.duration
+      - displayName: Secret Name
+        path: pkiConfg.rootCertificateAuthority.secretName
+      - displayName: Server Certificate
+        path: pkiConfg.serverCertificate
+      - displayName: Duration
+        path: pkiConfg.serverCertificate.duration
+      - displayName: Secret Name
+        path: pkiConfg.serverCertificate.secretName
+      - description: Resources holds the Resource Requirements to use for the discovery service Deployment. When not set it defaults to no resource requests nor limits. CPU and Memory resources are supported.
+        displayName: Resources
+        path: resources
+      - description: XdsServerPort is the port where the xDS server listens. Defaults to 18000.
+        displayName: Xds Server Port
+        path: xdsPort
+      statusDescriptors:
+      - description: Conditions represent the latest available observations of an object's state
+        displayName: Conditions
+        path: conditions
+      version: v1alpha1
+    - description: DiscoveryServiceCertificate is used to create certificates, either self-signed or by using a cert-manager CA issuer. This object is used by the DiscoveryService controller to create the required certificates for the different components of the discovery service. Direct use of DiscoveryServiceCertificate objects is discouraged.
+      displayName: DiscoveryServiceCertificate
+      kind: DiscoveryServiceCertificate
+      name: discoveryservicecertificates.operator.marin3r.3scale.net
+      specDescriptors:
+      - description: CertificateRenewalConfig configures the certificate renewal process. If unset default behavior is to renew the certificate but not notify of renewals.
+        displayName: Certificate Renewal Config
+        path: certificateRenewalNotification
+      - description: Enabled is a flag to enable or disable renewal of the certificate
+        displayName: Enabled
+        path: certificateRenewalNotification.enabled
+      - description: CommonName is the CommonName of the certificate
+        displayName: Common Name
+        path: commonName
+      - description: Hosts is the list of hosts the certificate is valid for. Only use when 'IsServerCertificate' is true. If unset, the CommonName field will be used to populate the valid hosts of the certificate.
+        displayName: Hosts
+        path: hosts
+      - description: IsCA is a boolean specifying that the certificate is a CA
+        displayName: Is CA
+        path: isCA
+      - description: SecretRef is a reference to the secret that will hold the certificate and the private key.
+        displayName: Secret Ref
+        path: secretRef
+      - description: IsServerCertificate is a boolean specifying if the certificate should be issued with server auth usage enabled
+        displayName: Is Server Certificate
+        path: server
+      - description: Signer specifies  the signer to use to create this certificate. Supported signers are CertManager and SelfSigned.
+        displayName: Signer
+        path: signer
+      - description: CASigned holds specific configuration for the CASigned signer
+        displayName: CASigned
+        path: signer.caSigned
+      - description: A reference to a Secret containing the CA
+        displayName: Secret Ref
+        path: signer.caSigned.caSecretRef
+      - description: SelfSigned holds specific configuration for the SelfSigned signer
+        displayName: Self Signed
+        path: signer.selfSigned
+      - description: ValidFor specifies the validity of the certificate in seconds
+        displayName: Valid For
+        path: validFor
+      statusDescriptors:
+      - description: Conditions represent the latest available observations of an object's state
+        displayName: Conditions
+        path: conditions
+      version: v1alpha1
   description: |
     marin3r is an operator to deploy and manage an envoy discovery service. It's main purpose is to provide
     an easy way to deploy a discovery service with the basic functionality already in place so you
