+++ controllers/operator.marin3r/ca_signed_test.go
@@ -4,7 +4,7 @@ import (
 	"crypto/x509"
 	"testing"
 
-	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator.marin3r/v1alpha1"
+	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator/v1alpha1"
 	"github.com/3scale/marin3r/pkg/util/pki"
 	v1 "k8s.io/api/core/v1"
 )
