+++ controllers/operator.marin3r/discoveryservice_controller.go
@@ -21,7 +21,7 @@ import (
 	"time"
 
 	marin3rv1alpha1 "github.com/3scale/marin3r/apis/marin3r/v1alpha1"
-	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator.marin3r/v1alpha1"
+	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator/v1alpha1"
 
 	"github.com/3scale/marin3r/pkg/reconcilers"
 	"github.com/go-logr/logr"
