+++ controllers/operator.marin3r/envoy_bootstrap.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	marin3rv1alpha1 "github.com/3scale/marin3r/apis/marin3r/v1alpha1"
-	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator.marin3r/v1alpha1"
+	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator/v1alpha1"
 	"github.com/3scale/marin3r/pkg/webhooks/podv1mutator"
 	"github.com/go-logr/logr"
 	"k8s.io/apimachinery/pkg/api/errors"
