+++ controllers/operator.marin3r/self_signed_test.go
@@ -3,7 +3,7 @@ package controllers
 import (
 	"testing"
 
-	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator.marin3r/v1alpha1"
+	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator/v1alpha1"
 )
 
 func Test_genSelfSignedCertificateObject(t *testing.T) {
