+++ examples/e2e/deployment/discoveryservice.yaml
@@ -2,10 +2,9 @@ apiVersion: operator.marin3r.3scale.net/v1alpha1
 kind: DiscoveryService
 metadata:
   name: instance
+  namespace: default
 spec:
   image: quay.io/3scale/marin3r:test
-  discoveryServiceNamespace: default
   debug: true
-  enabledNamespaces:
-    - default
+
