+++ examples/e2e/deployment/kuard.yaml
@@ -3,6 +3,7 @@ apiVersion: apps/v1
 kind: Deployment
 metadata:
   name: kuard
+  namespace: default
   labels:
     app: kuard
 spec:
