+++ pkg/reconcilers/marin3r/bootstrap_config.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 
 	marin3rv1alpha1 "github.com/3scale/marin3r/apis/marin3r/v1alpha1"
-	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator.marin3r/v1alpha1"
+	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator/v1alpha1"
 	"github.com/3scale/marin3r/pkg/envoy"
 	envoy_bootstrap "github.com/3scale/marin3r/pkg/envoy/bootstrap"
 	envoy_bootstrap_options "github.com/3scale/marin3r/pkg/envoy/bootstrap/options"
