+++ pkg/reconcilers/marin3r/bootstrap_config_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	marin3rv1alpha1 "github.com/3scale/marin3r/apis/marin3r/v1alpha1"
-	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator.marin3r/v1alpha1"
+	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator/v1alpha1"
 	"github.com/3scale/marin3r/pkg/envoy"
 	"github.com/go-logr/logr"
 	corev1 "k8s.io/api/core/v1"
