+++ test/e2e/marin3r/pod_suite_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/3scale/marin3r/apis/marin3r/v1alpha1"
 	marin3rv1alpha1 "github.com/3scale/marin3r/apis/marin3r/v1alpha1"
-	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator.marin3r/v1alpha1"
+	operatorv1alpha1 "github.com/3scale/marin3r/apis/operator/v1alpha1"
 	"github.com/3scale/marin3r/pkg/envoy"
 	testutil "github.com/3scale/marin3r/test/e2e/util"
 	. "github.com/onsi/ginkgo"
